Adds option to pass list to Neighborhood
Adds a few lines to `porchlight.neighborhood.__init__` to handle a new
kwarg, `initial_doors`, which is a list of `Door`/function objects.

These can be mixed! They do not all have to be doors or all have to be
functions, it just iterated through the list and uses `add_door` and
`add_function` methods depending on the object.