Adding flag for a basic auto-wrapping functionality
This commit includes a significant update to `Door`, introducing the
concept of "auto-wrapping" functions by passing a wrapped flag alongside
the expected parameters.

This commit also fixes a significant issue (which is addressed in the
issue and will be brought up in the relevant pull request) with
type-checking being on by default whenever using Door. This hadn't
caused issues in the past because ~~my code is always perfect~~ because
it wasn't implicitly covered in any tests.

This means 0.4.0 will not be fully backwards-compatible.