Updating return value handling
This commit updates the `return_vals` `BaseDoor` attr to be a list, and
for all `BaseDoor` objects to assert that return instances are unique
across all potential function returns.

This guarantee does not cover automatic functions, and it may be prudent
to add a warning for that in the future if return values don't map
clearly to expectations.