yield statements now supported
Changed one regex statement to include `yield` alongside `return` as an
option.

Also added a test to account for a yield statement, though this test
isn't particularly thorough. Given to `porchlight`, this really is just
a change in return vs. yield (since the user will still need to set this
up so `next` is called on the generator).

A future update could include more explicit generator support, but I'm
not sure that's feasible for 1.0.