Added `empty` filter to default dict
Initialization was providing default values for required arguments, in
the form of the `param.Empty` object, which will not necessarily cause
initialization to fail. It now filters out those positional arguments
that are also techincally keyword arguments.

In a future major update, it would be good to have a `Door` method that
retrieves these default values.