Initial commit of side-effect checking for Doors
This adds `Door._find_possible_side_effects` (a static method, which may
be better organized in `utils.py` eventually), and
`Door.side_effects_dict`, a property that returns a dictionary with the
line numbers and accompanying lines (as strs) that may cause side
effects.

This is a purely diagnostic tool; it does nothing more than look for the
most common Python patterns that result in side effects.