Add listeners to Param
This adds listeners to Param objects, which are called and passed a
reference to the notifying Param object (and nothing else).

This is mainly targeted at neighorhoods, which will implement a
Neighborhood.notify_listeners method that accepts a param and notifies
relevant doors of the change (whether they are in the main run order or
not).